отрефакторил попап
