+++ src/utils/common.js
@@ -1,7 +1,3 @@
-import dayjs from 'dayjs';
-const relativeTime = require('dayjs/plugin/relativeTime');
-dayjs.extend(relativeTime);
-
 const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -23,24 +19,16 @@ const getRandomArrayElements = (arr, amount = 1) => {
   return result;
 };
 
-const getYear = (date) => date ? dayjs(date).format('YYYY') : '';
-
-const getHumanDate = (date) => date ? dayjs(date).format('DD MMMM YYYY') : '';
-
-const getHumanDateTime = (date) => date ? dayjs(date).format('YYYY/MM/DD hh:mm') : '';
-
-const getHumanRelativeTime = (date) => date ? dayjs(date).fromNow() : '';
-
 const findIndexByValue = (arr, value) => {
   const index = arr.findIndex((element) => element === value);
   return index;
-}
+};
 
-const removeIndexFromArray = (arr, index) => {
-  return [
+const removeIndexFromArray = (arr, index) => (
+  [
     ...arr.slice(0, index),
     ...arr.slice(index + 1),
-  ];
-}
+  ]
+);
 
-export {removeIndexFromArray, findIndexByValue, getRandomInteger, getRandomDecimal, getRandomArrayElement, getRandomArrayElements, getYear, getHumanDate, getHumanDateTime, getHumanRelativeTime};
+export {removeIndexFromArray, findIndexByValue, getRandomInteger, getRandomDecimal, getRandomArrayElement, getRandomArrayElements};
