+++ src/view/filter.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { FilterType, FilterName } from '../const.js';
+import { FilterType } from '../const.js';
 
 const createFilterItemTemplate  = (filter, isActive) => {
   const {name, displayableName, count} = filter;
